cleanup

+++ src/predictor/loss_score.py
@@ -20,7 +20,7 @@ class LossScore(nn.Module):
         # Higher loss for misprediction of low rated items
         is_3_or_less_stars_penalty = (target <= 0.5).float() * 0.6
         is_5_penalty = (target >= 0.99).float() * 0.1
-        return (loss + is_3_or_less_stars_penalty.float() + is_5_penalty.float()).mean()
+        return (loss + is_3_or_less_stars_penalty + is_5_penalty).mean()
 
     def MSE_variation1(self, output, target):
         loss = torch.square(target - output)
